+++ src/spimdisasm/src/sections/python_bindings/py_data_section.rs
@@ -25,7 +25,10 @@ enum PyDataSectionInner {
 }
 
 #[derive(Debug, Clone, Hash, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "pyo3", pyclass(module = "spimdisasm", name = "n", from_py_object))]
+#[cfg_attr(
+    feature = "pyo3",
+    pyclass(module = "spimdisasm", name = "DataSection", from_py_object)
+)]
 pub struct PyDataSection {
     inner: PyDataSectionInner,
 }