+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change global segments to allow more than one global segment instead of
   forcing the user to merge all the ranges into one.
+- Move the `GlobalOffsetTable`, `GotGlobalEntry`, `GotLocalEntry`,
+  `GotRequestedAddress` and `GlobalOffsetTableIter` to the new `got` module from
+  the `addresses` module.
 
 ## [2.0.0-alpha.1] - 2025-08-07
 
@@ -1703,6 +1706,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version 1.0.0
 
 [unreleased]: https://github.com/Decompollaborate/spimdisasm/compare/2.0.0-alpha.1...rs
+
 [2.0.0-alpha.1]: https://github.com/Decompollaborate/spimdisasm/compare/1.30.2...2.0.0-alpha.1
 [1.30.2]: https://github.com/Decompollaborate/spimdisasm/compare/1.30.1...1.30.2
 [1.30.1]: https://github.com/Decompollaborate/spimdisasm/compare/1.30.0...1.30.1